Email, 기타 설정 수정
