+++ src/main/java/com/fromits/controller/PromiseController.java
@@ -11,11 +11,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-import java.util.HashMap;
-import java.util.List;
+import javax.mail.*;
+import javax.mail.internet.InternetAddress;
+import javax.mail.internet.MimeMessage;
+import java.util.*;
 
 import java.util.List;
-import java.util.Map;
 
 @Controller
 @Slf4j
@@ -29,10 +30,47 @@ public class PromiseController {
     final GroupmemberService groupmemberService;
     final DevoteService devoteService;
     final CustService custService;
+
+    @Value("${app.key.username}")
+    String mail;
+    @Value("${app.key.password}")
+    String pwd;
+    public static void sendEmail(String to, String from, String host, String subject, String text, String pwd) {
+        // 메일 서버 연결을 위한 프로퍼티 설정
+        Properties properties = System.getProperties();
+        properties.setProperty("mail.smtp.host", "smtp.gmail.com");
+        properties.setProperty("mail.smtp.port", "587"); // 대부분의 SMTP 서버는 587 포트 사용
+        properties.setProperty("mail.smtp.auth", "true");
+        properties.setProperty("mail.smtp.starttls.enable", "true"); // TLS 사용 설정
+
+        // 인증 정보
+        Session session = Session.getInstance(properties, new Authenticator() {
+            protected PasswordAuthentication getPasswordAuthentication() {
+
+                return new PasswordAuthentication(from, pwd); // 발신자 이메일 계정과 비밀번호 입력
+            }
+        });
+
+        try {
+            // 메일 생성
+            MimeMessage message = new MimeMessage(session);
+            message.setFrom(new InternetAddress(from));
+            message.addRecipient(Message.RecipientType.TO, new InternetAddress(to));
+            message.setSubject(subject);
+            message.setText(text);
+
+            // 메일 전송
+            Transport.send(message);
+            System.out.println("Sent message successfully....");
+        } catch (MessagingException mex) {
+            mex.printStackTrace();
+        }
+
+
 //
-//
-//    @Value("${app.key.kakaokey}")
-//    private String kakaoKey;
+
+    }
+
 
     @RequestMapping("/schedule")
     public String promise(Model model) throws Exception {
@@ -82,7 +120,6 @@ public class PromiseController {
 
     @RequestMapping("/finalpromise")
     public String finalpromise(Model model, @RequestParam("devoteId") int devoteId) throws Exception {
-
         DevoteDto devoteDto = devoteService.getProId(devoteId);
         Map<String, Integer> placeresult = new HashMap<>();
         FinalPlaceDto finalPlace = promiseService.finalplace(devoteId);
@@ -138,9 +175,12 @@ public class PromiseController {
         Map<String, Object> params = new HashMap<>();
         params.put("proDate", proDate);
         params.put("proId", proId);
+        List<String> emails = promiseService.getEmailsbyProId(Integer.parseInt(proId));
+        for(String s: emails){
+            sendEmail(s, mail, "smtp.gmail.com", "From-its", "투표가 완료되었습니다. 확인하세요", pwd);
+        }
 
-        int result = promiseService.finalPromiseSchedule(
-                params);
+        int result = promiseService.finalPromiseSchedule(params);
 
         return result;
     }
