+++ src/main/resources/application-dev.yml
@@ -31,12 +31,9 @@ spring:
 
   datasource:
     driver-class-name: net.sf.log4jdbc.sql.jdbcapi.DriverSpy
-    url: jdbc:log4jdbc:mysql://127.0.0.1:3306/fromits?characterEncoding=utf8
-    #    driver-class-name: com.mysql.cj.jdbc.Driver
-    #    url: jdbc:mysql://127.0.0.1:3306/shopdb?characterEncoding=utf8&serverTimezone=Asia/Seoul
-    username: fromuser
-    # 각자의 id, password를 입력해야됩니다!
-    password: 111111
+    url: jdbc:log4jdbc:mysql://fromits.cl82qi2c2rep.ap-northeast-2.rds.amazonaws.com/fromits?characterEncoding=utf8
+    username: ENC(Nh56kHVt0DqOblRXYvnJaw==)
+    password: ENC(MKkltWk225tIjbuRxlV7kgea59Z1bRG1)
 mybatis:
   type-aliases-package: com.fromits.app.dto
   mapper-locations: classpath:/mapper/*.xml
